Create a data.frame with documentation metadata of one or more packages

Create a data.frame with documentation metadata of one or more packages. If you develop a universe, such as the tidyverse or tidymodels, dverse helps you to easily create a universe-wide reference for the pkgdown website of your meta-package.

Usage

document_universe(
  x,
  url_template = NULL,
  url = NULL,
  packages = NULL,
  strip_s3class = TRUE
)

Arguments

x

A character vector giving concepts or package names to match.

url_template

Character. A template to generate links to documentation based on the column names of the output – typically package and topic, e.g. "https://maurolepore.github.io/{package}/reference/{topic}.html" (glue::glue() syntax).

url

Character vector of length-1 giving a base url, e.g. "https://maurolepore.github.io/".

packages

A character vector with the names of packages to limit the search, or NULL to serach in all known libraries.

strip_s3class

TRUE removes the class component of S3 methods.

Value

A dataframe.

Examples

document_universe("datasets")
#>                    topic
#> 1          AirPassengers
#> 2                BJsales
#> 3                    BOD
#> 4                    CO2
#> 5            ChickWeight
#> 6                  DNase
#> 7         EuStockMarkets
#> 8           Formaldehyde
#> 9           HairEyeColor
#> 10          Harman23.cor
#> 11          Harman74.cor
#> 12              Indometh
#> 13          InsectSprays
#> 14        JohnsonJohnson
#> 15             LakeHuron
#> 16      LifeCycleSavings
#> 17              Loblolly
#> 18                  Nile
#> 19                Orange
#> 20         OrchardSprays
#> 21           PlantGrowth
#> 22             Puromycin
#> 23                Theoph
#> 24               Titanic
#> 25           ToothGrowth
#> 26         UCBAdmissions
#> 27        UKDriverDeaths
#> 28          UKLungDeaths
#> 29                 UKgas
#> 30           USAccDeaths
#> 31             USArrests
#> 32        USJudgeRatings
#> 33 USPersonalExpenditure
#> 34              VADeaths
#> 35              WWWusage
#> 36           WorldPhones
#> 37           ability.cov
#> 38              airmiles
#> 39            airquality
#> 40              anscombe
#> 41                attenu
#> 42              attitude
#> 43               austres
#> 44               beavers
#> 45                  cars
#> 46              chickwts
#> 47                   co2
#> 48               crimtab
#> 49      datasets-package
#> 50           discoveries
#> 51                 esoph
#> 52                  euro
#> 53              eurodist
#> 54              faithful
#> 55                freeny
#> 56                infert
#> 57                  iris
#> 58               islands
#> 59                    lh
#> 60               longley
#> 61                  lynx
#> 62                morley
#> 63                mtcars
#> 64                nhtemp
#> 65                nottem
#> 66                   npk
#> 67    occupationalStatus
#> 68                precip
#> 69            presidents
#> 70              pressure
#> 71                quakes
#> 72                 randu
#> 73                rivers
#> 74                  rock
#> 75                 sleep
#> 76             stackloss
#> 77                 state
#> 78         sunspot.month
#> 79          sunspot.year
#> 80              sunspots
#> 81                 swiss
#> 82              treering
#> 83                 trees
#> 84                 uspop
#> 85               volcano
#> 86            warpbreaks
#> 87                 women
#>                                                                                              alias
#> 1                                                                                    AirPassengers
#> 2                                                                            BJsales, BJsales.lead
#> 3                                                                                              BOD
#> 4                                                                                              CO2
#> 5                                                                                      ChickWeight
#> 6                                                                                            DNase
#> 7                                                                                   EuStockMarkets
#> 8                                                                                     Formaldehyde
#> 9                                                                                     HairEyeColor
#> 10                                                                                    Harman23.cor
#> 11                                                                                    Harman74.cor
#> 12                                                                                        Indometh
#> 13                                                                                    InsectSprays
#> 14                                                                                  JohnsonJohnson
#> 15                                                                                       LakeHuron
#> 16                                                                                LifeCycleSavings
#> 17                                                                                        Loblolly
#> 18                                                                                            Nile
#> 19                                                                                          Orange
#> 20                                                                                   OrchardSprays
#> 21                                                                                     PlantGrowth
#> 22                                                                                       Puromycin
#> 23                                                                                          Theoph
#> 24                                                                                         Titanic
#> 25                                                                                     ToothGrowth
#> 26                                                                                   UCBAdmissions
#> 27                                                                       UKDriverDeaths, Seatbelts
#> 28                                                         UKLungDeaths, ldeaths, fdeaths, mdeaths
#> 29                                                                                           UKgas
#> 30                                                                                     USAccDeaths
#> 31                                                                                       USArrests
#> 32                                                                                  USJudgeRatings
#> 33                                                                           USPersonalExpenditure
#> 34                                                                                        VADeaths
#> 35                                                                                        WWWusage
#> 36                                                                                     WorldPhones
#> 37                                                                                     ability.cov
#> 38                                                                                        airmiles
#> 39                                                                                      airquality
#> 40                                                                                        anscombe
#> 41                                                                                          attenu
#> 42                                                                                        attitude
#> 43                                                                                         austres
#> 44                                                                       beavers, beaver1, beaver2
#> 45                                                                                            cars
#> 46                                                                                        chickwts
#> 47                                                                                             co2
#> 48                                                                                         crimtab
#> 49                                                                      datasets-package, datasets
#> 50                                                                                     discoveries
#> 51                                                                                           esoph
#> 52                                                                                euro, euro.cross
#> 53                                                                             eurodist, UScitiesD
#> 54                                                                                        faithful
#> 55                                                                      freeny, freeny.x, freeny.y
#> 56                                                                                          infert
#> 57                                                                                     iris, iris3
#> 58                                                                                         islands
#> 59                                                                                              lh
#> 60                                                                                         longley
#> 61                                                                                            lynx
#> 62                                                                                          morley
#> 63                                                                                          mtcars
#> 64                                                                                          nhtemp
#> 65                                                                                          nottem
#> 66                                                                                             npk
#> 67                                                                              occupationalStatus
#> 68                                                                                          precip
#> 69                                                                                      presidents
#> 70                                                                                        pressure
#> 71                                                                                          quakes
#> 72                                                                                           randu
#> 73                                                                                          rivers
#> 74                                                                                            rock
#> 75                                                                                           sleep
#> 76                                                                  stackloss, stack.loss, stack.x
#> 77 state, state.abb, state.area, state.center, state.division, state.name, state.region, state.x77
#> 78                                                                                   sunspot.month
#> 79                                                                                    sunspot.year
#> 80                                                                                        sunspots
#> 81                                                                                           swiss
#> 82                                                                                        treering
#> 83                                                                                           trees
#> 84                                                                                           uspop
#> 85                                                                                         volcano
#> 86                                                                                      warpbreaks
#> 87                                                                                           women
#>                                                              title
#> 1                      Monthly Airline Passenger Numbers 1949-1960
#> 2                                Sales Data with Leading Indicator
#> 3                                        Biochemical Oxygen Demand
#> 4                            Carbon Dioxide Uptake in Grass Plants
#> 5                   Weight versus age of chicks on different diets
#> 6                                             Elisa assay of DNase
#> 7  Daily Closing Prices of Major European Stock Indices, 1991-1998
#> 8                                    Determination of Formaldehyde
#> 9                        Hair and Eye Color of Statistics Students
#> 10                                              Harman Example 2.3
#> 11                                              Harman Example 7.4
#> 12                                Pharmacokinetics of Indomethacin
#> 13                                  Effectiveness of Insect Sprays
#> 14                  Quarterly Earnings per Johnson & Johnson Share
#> 15                                   Level of Lake Huron 1875-1972
#> 16                            Intercountry Life-Cycle Savings Data
#> 17                                   Growth of Loblolly Pine Trees
#> 18                                          Flow of the River Nile
#> 19                                          Growth of Orange Trees
#> 20                                       Potency of Orchard Sprays
#> 21                      Results from an Experiment on Plant Growth
#> 22                      Reaction Velocity of an Enzymatic Reaction
#> 23                                Pharmacokinetics of Theophylline
#> 24                           Survival of passengers on the Titanic
#> 25          The Effect of Vitamin C on Tooth Growth in Guinea Pigs
#> 26                               Student Admissions at UC Berkeley
#> 27                        Road Casualties in Great Britain 1969-84
#> 28                     Monthly Deaths from Lung Diseases in the UK
#> 29                                    UK Quarterly Gas Consumption
#> 30                           Accidental Deaths in the US 1973-1978
#> 31                                 Violent Crime Rates by US State
#> 32       Lawyers' Ratings of State Judges in the US Superior Court
#> 33                                       Personal Expenditure Data
#> 34                                  Death Rates in Virginia (1940)
#> 35                                       Internet Usage per Minute
#> 36                                          The World's Telephones
#> 37                                  Ability and Intelligence Tests
#> 38            Passenger Miles on Commercial US Airlines, 1937-1960
#> 39                               New York Air Quality Measurements
#> 40     Anscombe's Quartet of 'Identical' Simple Linear Regressions
#> 41                               The Joyner-Boore Attenuation Data
#> 42                              The Chatterjee-Price Attitude Data
#> 43     Quarterly Time Series of the Number of Australian Residents
#> 44                          Body Temperature Series of Two Beavers
#> 45                            Speed and Stopping Distances of Cars
#> 46                                    Chicken Weights by Feed Type
#> 47                         Mauna Loa Atmospheric CO2 Concentration
#> 48                                   Student's 3000 Criminals Data
#> 49                                          The R Datasets Package
#> 50                         Yearly Numbers of Important Discoveries
#> 51                       Smoking, Alcohol and (O)esophageal Cancer
#> 52                             Conversion Rates of Euro Currencies
#> 53         Distances Between European Cities and Between US Cities
#> 54                                        Old Faithful Geyser Data
#> 55                                           Freeny's Revenue Data
#> 56              Infertility after Spontaneous and Induced Abortion
#> 57                                      Edgar Anderson's Iris Data
#> 58                           Areas of the World's Major Landmasses
#> 59                            Luteinizing Hormone in Blood Samples
#> 60                              Longley's Economic Regression Data
#> 61                        Annual Canadian Lynx trappings 1821-1934
#> 62                                   Michelson Speed of Light Data
#> 63                                      Motor Trend Car Road Tests
#> 64                        Average Yearly Temperatures in New Haven
#> 65           Average Monthly Temperatures at Nottingham, 1920-1939
#> 66                          Classical N, P, K Factorial Experiment
#> 67                   Occupational Status of Fathers and their Sons
#> 68                               Annual Precipitation in US Cities
#> 69                     Quarterly Approval Ratings of US Presidents
#> 70          Vapor Pressure of Mercury as a Function of Temperature
#> 71                               Locations of Earthquakes off Fiji
#> 72                Random Numbers from Congruential Generator RANDU
#> 73                          Lengths of Major North American Rivers
#> 74                          Measurements on Petroleum Rock Samples
#> 75                                            Student's Sleep Data
#> 76                                Brownlee's Stack Loss Plant Data
#> 77                                      US State Facts and Figures
#> 78                    Monthly Sunspot Data, from 1749 to "Present"
#> 79                                  Yearly Sunspot Data, 1700-1988
#> 80                              Monthly Sunspot Numbers, 1749-1983
#> 81        Swiss Fertility and Socioeconomic Indicators (1888) Data
#> 82                               Yearly Tree-Ring Data, -6000-1979
#> 83              Diameter, Height and Volume for Black Cherry Trees
#> 84                           Populations Recorded by the US Census
#> 85       Topographic Information on Auckland's Maunga Whau Volcano
#> 86                     The Number of Breaks in Yarn during Weaving
#> 87                  Average Heights and Weights for American Women
#>                         concept  package
#> 1  Datasets available by data() datasets
#> 2  Datasets available by data() datasets
#> 3  Datasets available by data() datasets
#> 4  Datasets available by data() datasets
#> 5  Datasets available by data() datasets
#> 6  Datasets available by data() datasets
#> 7  Datasets available by data() datasets
#> 8  Datasets available by data() datasets
#> 9  Datasets available by data() datasets
#> 10 Datasets available by data() datasets
#> 11 Datasets available by data() datasets
#> 12 Datasets available by data() datasets
#> 13 Datasets available by data() datasets
#> 14 Datasets available by data() datasets
#> 15 Datasets available by data() datasets
#> 16 Datasets available by data() datasets
#> 17 Datasets available by data() datasets
#> 18 Datasets available by data() datasets
#> 19 Datasets available by data() datasets
#> 20 Datasets available by data() datasets
#> 21 Datasets available by data() datasets
#> 22 Datasets available by data() datasets
#> 23 Datasets available by data() datasets
#> 24 Datasets available by data() datasets
#> 25 Datasets available by data() datasets
#> 26 Datasets available by data() datasets
#> 27 Datasets available by data() datasets
#> 28 Datasets available by data() datasets
#> 29 Datasets available by data() datasets
#> 30 Datasets available by data() datasets
#> 31 Datasets available by data() datasets
#> 32 Datasets available by data() datasets
#> 33 Datasets available by data() datasets
#> 34 Datasets available by data() datasets
#> 35 Datasets available by data() datasets
#> 36 Datasets available by data() datasets
#> 37 Datasets available by data() datasets
#> 38 Datasets available by data() datasets
#> 39 Datasets available by data() datasets
#> 40 Datasets available by data() datasets
#> 41 Datasets available by data() datasets
#> 42 Datasets available by data() datasets
#> 43 Datasets available by data() datasets
#> 44 Datasets available by data() datasets
#> 45 Datasets available by data() datasets
#> 46 Datasets available by data() datasets
#> 47 Datasets available by data() datasets
#> 48 Datasets available by data() datasets
#> 49            Package Summaries datasets
#> 50 Datasets available by data() datasets
#> 51 Datasets available by data() datasets
#> 52 Datasets available by data() datasets
#> 53 Datasets available by data() datasets
#> 54 Datasets available by data() datasets
#> 55 Datasets available by data() datasets
#> 56 Datasets available by data() datasets
#> 57 Datasets available by data() datasets
#> 58 Datasets available by data() datasets
#> 59 Datasets available by data() datasets
#> 60 Datasets available by data() datasets
#> 61 Datasets available by data() datasets
#> 62 Datasets available by data() datasets
#> 63 Datasets available by data() datasets
#> 64 Datasets available by data() datasets
#> 65 Datasets available by data() datasets
#> 66 Datasets available by data() datasets
#> 67 Datasets available by data() datasets
#> 68 Datasets available by data() datasets
#> 69 Datasets available by data() datasets
#> 70 Datasets available by data() datasets
#> 71 Datasets available by data() datasets
#> 72 Datasets available by data() datasets
#> 73 Datasets available by data() datasets
#> 74 Datasets available by data() datasets
#> 75 Datasets available by data() datasets
#> 76 Datasets available by data() datasets
#> 77 Datasets available by data() datasets
#> 78 Datasets available by data() datasets
#> 79 Datasets available by data() datasets
#> 80 Datasets available by data() datasets
#> 81 Datasets available by data() datasets
#> 82 Datasets available by data() datasets
#> 83 Datasets available by data() datasets
#> 84 Datasets available by data() datasets
#> 85 Datasets available by data() datasets
#> 86 Datasets available by data() datasets
#> 87 Datasets available by data() datasets